pyproject.toml
[project]
name = "ate-api"
version = "0.6.0"
description = "Operational data API for Active Travel England."
requires-python = ">=3.13"
dependencies = [
    "asyncpg~=0.30.0",
    "authlib~=1.6.0",
    "fastapi[standard]~=0.115.0",
    "pydantic-settings~=2.10.0",
    "sqlalchemy[asyncio]~=2.0.0"
]

[project.optional-dependencies]
dev = [
    "black~=25.1.0",
    "isort~=6.0.0",
    "mypy~=1.16.0",
    "pytest~=8.4.0",
    "pytest-asyncio~=1.0.0",
    "respx~=0.22.0",
    "ruff~=0.12.0",
    "testcontainers~=4.10.0"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

# black

[tool.black]
line-length = 120

# isort

[tool.isort]
profile = "black"
line_length = 120

# mypy

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "authlib.*",
    "testcontainers.postgres.*"
]
ignore_missing_imports = true

# pytest

[tool.pytest.ini_options]
asyncio_mode = "auto"

# ruff

[tool.ruff]
line-length = 120
